+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/events/ServerEvents.java
@@ -5,13 +5,13 @@ import net.mehvahdjukaar.moonlight.api.events.IFireConsumeBlockEvent;
 import net.mehvahdjukaar.moonlight.api.fluids.BuiltInSoftFluids;
 import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidRegistry;
 import net.mehvahdjukaar.moonlight.api.misc.EventCalled;
+import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
 import net.mehvahdjukaar.supplementaries.SuppPlatformStuff;
 import net.mehvahdjukaar.supplementaries.api.IQuiverEntity;
 import net.mehvahdjukaar.supplementaries.common.block.IRopeConnection;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.AshLayerBlock;
 import net.mehvahdjukaar.supplementaries.common.block.faucet.FaucetBehaviorsManager;
 import net.mehvahdjukaar.supplementaries.common.block.hourglass.HourglassTimesManager;
-import net.mehvahdjukaar.supplementaries.common.block.tiles.EndermanSkullBlockTile;
 import net.mehvahdjukaar.supplementaries.common.entities.ISlimeable;
 import net.mehvahdjukaar.supplementaries.common.entities.goals.EatFodderGoal;
 import net.mehvahdjukaar.supplementaries.common.entities.goals.EvokerRedMerchantWololooSpellGoal;
@@ -34,8 +34,8 @@ import net.mehvahdjukaar.supplementaries.reg.ModSetup;
 import net.mehvahdjukaar.supplementaries.reg.ModTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.core.RegistryAccess;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.stats.Stats;
@@ -157,7 +157,7 @@ public class ServerEvents {
 
     @EventCalled
     public static void onServerStart(MinecraftServer server) {
-        FaucetBehaviorsManager.RELOAD_INSTANCE.onLevelLoad(server.overworld());
+        FaucetBehaviorsManager.RELOAD_INSTANCE.onReloadWithLevel(server.overworld());
     }
 
     @EventCalled
@@ -175,6 +175,13 @@ public class ServerEvents {
             }
         }
         CurseLootFunction.rebuild();
+
+        if (!client) {
+            var server = PlatHelper.getCurrentServer();
+            if (server != null) {
+                FaucetBehaviorsManager.RELOAD_INSTANCE.onReloadWithLevel(server.overworld());
+            }
+        }
     }
 
     private static final boolean FODDER_ENABLED = CommonConfigs.Functional.FODDER_ENABLED.get();