+++ changelog.md
@@ -1,3 +1,5 @@
-CHECK 3.0.0 PATCH NOTES!!
-
-- No more joking missing texture. If mods are screwing up and force setting my crop to invalid values they will now get a nice texture blaming them
+- fixed faucet brewing stand iteraction
+- updated integration with buzzier bees
+- improved gunpowder connections
+- fixed faucet emptying fluidlogged blocks incorrectly
+- fixed faucet stopping working after datapack reload